Fix score (should be better values now).
runCommand() now redirects all stderr output to stdout. This will help redirecting the scripts output to a file if the user wanted to save the results and compare.

Some programs that are going to be passed to runCommand() will write some data to stderr, which can be text we don't want. This sends that data to stdout instead.