Allow testing without checking the command for certain programs.
Extraction will still count compression, which will be wrong, but I think score should be ok, but it could also be wrong.

So, score could work but compression especially should always be ignored when running a command that results in decompression.

Resulting in determining what mode the arguments provide is too much work. Sticking with an argument for a input path and a output path will always work regardless of the command we run.